Resolved null value issue with getRank()
where the application would crash if a name was not found for a given
year and the method would return a null value. Revised method so that a
zero is returned when a name is not found in a given year. Tested and
application works as intended

Added comments to increase clarity and readability